requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-dotenv
aio_pika
asyncpg
pytest
httpx
